Factor tool output handlers into dedicated modules
